pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "reuse-license-utils"
description = "A collection of utilities for managing licensing in projects compliant with REUSE"
requires-python = ">= 3.10"
readme = "README.md"
dependencies = [
    "reuse~=6.0",
    "pydantic",
    "tomli; python_version<'3.11'",
    "gitpython~=3.1",
]
dynamic = ["version"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "reuse_license_utils/version.py"

[dependency-groups]
format = ["ruff"]
dev = [{ include-group="format" }]

[tool.uv]
default-groups = ["dev"]
package = true

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.10
target-version = "py310"
fix = false
namespace-packages = ["a4x/orchestration"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint]
select = [
    "D",
    "YTT",   # flake8-2020
    "ANN",   # flake8-annotations
    "ASYNC", # flake8-async
    "BLE",   # flake8-blind-except
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "COM",   # flake8-commas
    "C4",    # flake8-comprehensions
    "EXE",   # flake8-executable
    "FA",    # flake8-future-annotations
    "ICN",   # flake8-import-conventions
    "PIE",   # flake8-pie
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RET",   # flake8-return
    "SLF",   # flake8-self
    "TC",    # flake8-type-checking
    "PTH",   # flake8-use-pathlib
    "I",     # isort
    "N",     # pep8-naming
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "PLE",   # Other errors
    "PLR",   # Other refactors
    "PLW",   # Other warnings
    "RUF",   # Ruff-specific rules
    # TODO uncomment CPY when we start strictly enforcing copyright headers
    # "CPY",   # flake8-copyright
]
ignore = [
    "ANN002",
    "ANN003",
    "ANN401",
    "B011",
    "B019",
    "B028",
    "B903",
    "B905",
    "B911",
    "B912",
    "INP",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D205",
    "D212",
    "D400",
    # "D415",  # Uncomment to allow docstrings to not end in punctuation
    "E402",
    "RUF012",
    "PLR0913",
    "PLW0603",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
# Format code blocks in docstrings
docstring-code-format = true
